+++ extensions/better-deal/CHANGELOG.md
@@ -1,5 +1,10 @@
 # Better Deal Changelog
 
+## [New Features] - 2024-09-05
+
+- Add delete single record
+- Add delete all records
+
 ## [Initial Version] - 2024-08-24
 
 Initial version code